Reinvite sdp event can be completed asynchronously
If the user has registered a listener for SDP, when the SDP
is from a remote re-invite, we pass an extra parameter in the event
we forward to the user, e.await. Calling this function will return a
callback the user must call when they have finished processing the sdp.
This allows them to asynchronously process the SDP.
This is handy for starleaf when hacking a restart of the pc stream
to appease the temasys plugin